Refactoring:
- make SemanticAnalyzer a trait
- Rename Parser to ParserInterpreter
- include source scanning into ParserInterpreter
