+++ src/main/scala/cmd/App.scala
@@ -1,23 +1,16 @@
 package cmd
 
+import interpreter.ParserInterpreter
+
 import java.nio.charset.Charset
 import java.nio.file.{Files, Paths}
 
-import parser.Parser
-import scanner.Scanner
-import semantic.SemanticAnalyzer.SemanticError
-
 object App extends App {
   val options = Cli.parse(args)
+
   val sourcePath = options(Cli.Source).toString
-
   val source = new String(Files.readAllBytes(Paths.get(sourcePath)), Charset.defaultCharset)
-  val scanner = new Scanner(source)
-  val parser = new Parser(scanner.buffered)
 
-  try {
-    parser.parse()
-  } catch {
-    case _: SemanticError =>
-  }
+  val interpreter = new ParserInterpreter(source)
+  interpreter.run()
 }
