+++ src/main/scala/evaluator/Evaluator.scala
@@ -1,0 +1,7 @@
+package evaluator
+
+import parser.Parser
+
+trait Evaluator { this : Parser =>
+  def run(): Unit
+}
